Update manifest examples to show origin scheme
Currently the "https://" scheme is required by the `origin` fields in
`url_handlers`. This is consistent with Chromium's manifest parser
implementation. This change updates the manifest examples in this
explainer to reflect that. If there is feedback, we could consider
also allowing a short form of origin strings where the "https://" scheme
is implied.